+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -8,3 +8,4 @@
 * Adds Camera and UseCase classes, along with methods for binding UseCases to a lifecycle with the ProcessCameraProvider.
 * Bump CameraX version to 1.3.0-alpha03 and Kotlin version to 1.8.0.
 * Changes instance manager to allow the separate creation of identical objects.
+* Adds Preview and Surface classes, along with other methods needed to implement camera preview.